refactor(daemon): move wardnet-common inline tests to src/tests/
Move the inline #[cfg(test)] mod tests blocks out of dns.rs, net.rs and
serde_util.rs into the mandated src/tests/ layout (.agents/testing.md).
The dns.rs tests are merged into the pre-existing src/tests/dns.rs; net
and serde_util get new test files wired into src/tests/mod.rs. Pure
move — no test logic changed.

Part of #847.
