+++ source/daemon/crates/wardnet-common/src/dns.rs
@@ -502,151 +502,6 @@ impl DnsQueryResult {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn default_upstreams_include_quad9() {
-        let cfg = DnsConfig::default();
-        let addrs: Vec<&str> = cfg
-            .upstream_servers
-            .iter()
-            .map(|u| u.address.as_str())
-            .collect();
-        assert!(addrs.contains(&"1.1.1.1"), "Cloudflare present");
-        assert!(addrs.contains(&"8.8.8.8"), "Google present");
-        assert!(
-            addrs.contains(&"9.9.9.9"),
-            "Quad9 present on fresh installs"
-        );
-        let quad9 = cfg
-            .upstream_servers
-            .iter()
-            .find(|u| u.address == "9.9.9.9")
-            .expect("quad9");
-        assert_eq!(quad9.name, "Quad9");
-        assert_eq!(quad9.protocol, DnsProtocol::Udp);
-    }
-
-    #[test]
-    fn default_forwarder_selection_is_failover() {
-        let cfg = DnsConfig::default();
-        assert_eq!(
-            cfg.forwarder_selection_mode,
-            ForwarderSelectionMode::Failover
-        );
-        assert_eq!(cfg.single_upstream, None);
-    }
-
-    #[test]
-    fn forwarder_selection_mode_str_matches_serde() {
-        // as_str() must match the snake_case wire form used for KV persistence.
-        for m in [
-            ForwarderSelectionMode::Failover,
-            ForwarderSelectionMode::Fastest,
-            ForwarderSelectionMode::Single,
-        ] {
-            assert_eq!(
-                serde_json::to_string(&m).unwrap(),
-                format!("\"{}\"", m.as_str())
-            );
-            // from_wire round-trips through as_str.
-            assert_eq!(ForwarderSelectionMode::from_wire(m.as_str()), m);
-        }
-        // Unknown/legacy values fall back to the default.
-        assert_eq!(
-            ForwarderSelectionMode::from_wire("auto"),
-            ForwarderSelectionMode::Failover
-        );
-    }
-
-    #[test]
-    fn parse_all_resolver_strings() {
-        assert_eq!(
-            DnsQueryResult::parse("forwarded"),
-            DnsQueryResult::Forwarded
-        );
-        assert_eq!(DnsQueryResult::parse("cache_hit"), DnsQueryResult::CacheHit);
-        assert_eq!(DnsQueryResult::parse("blocked"), DnsQueryResult::Blocked);
-        assert_eq!(
-            DnsQueryResult::parse("blocked_skipped"),
-            DnsQueryResult::BlockedSkipped
-        );
-        assert_eq!(
-            DnsQueryResult::parse("rewritten"),
-            DnsQueryResult::Rewritten
-        );
-        assert_eq!(
-            DnsQueryResult::parse("recursive"),
-            DnsQueryResult::Recursive
-        );
-        assert_eq!(
-            DnsQueryResult::parse("upstream_error"),
-            DnsQueryResult::UpstreamError
-        );
-        assert_eq!(
-            DnsQueryResult::parse("authoritative"),
-            DnsQueryResult::Authoritative
-        );
-    }
-
-    #[test]
-    fn parse_unknown_falls_back_to_error() {
-        assert_eq!(DnsQueryResult::parse("gibberish"), DnsQueryResult::Error);
-        assert_eq!(DnsQueryResult::parse(""), DnsQueryResult::Error);
-        // Old aliases no longer accepted — the migration removes those rows.
-        assert_eq!(DnsQueryResult::parse("cached"), DnsQueryResult::Error);
-        assert_eq!(DnsQueryResult::parse("local"), DnsQueryResult::Error);
-    }
-
-    #[test]
-    fn as_str_round_trips_through_parse() {
-        for v in DnsQueryResult::ALL {
-            assert_eq!(DnsQueryResult::parse(v.as_str()), v);
-        }
-    }
-
-    /// `Error` round-trips through a real match arm, not the unknown-string
-    /// fallback.
-    ///
-    /// Without this, `as_str_round_trips_through_parse` passes vacuously for
-    /// `Error`: the fallback also returns `Error`, so a missing `"error"` arm
-    /// satisfies the assertion while silently routing every `"error"` row
-    /// through the "unknown DNS result string" path.
-    #[test]
-    fn every_variant_parses_without_hitting_the_fallback() {
-        for v in DnsQueryResult::ALL {
-            assert_eq!(
-                DnsQueryResult::from_db_str(v.as_str()),
-                Some(v),
-                "`{}` has no match arm — it would only survive the round-trip \
-                 via the unknown-string fallback",
-                v.as_str()
-            );
-        }
-    }
-
-    #[test]
-    fn from_db_str_rejects_unknown_strings() {
-        assert_eq!(DnsQueryResult::from_db_str("gibberish"), None);
-        assert_eq!(DnsQueryResult::from_db_str(""), None);
-    }
-
-    /// `ALL` really does hold every variant, exactly once.
-    ///
-    /// `slot()` is an exhaustive match, so a new variant cannot compile without
-    /// being given a slot; this asserts that slot lines up with a distinct
-    /// entry in `ALL`. Together they make `ALL` a guarantee rather than a
-    /// promise — every test that iterates it covers the new variant on day one.
-    #[test]
-    fn all_is_complete() {
-        for (i, v) in DnsQueryResult::ALL.into_iter().enumerate() {
-            assert_eq!(v.slot(), i, "ALL[{i}] = {v:?} is in the wrong slot");
-        }
-    }
-}
-
 /// A single entry in the DNS query log.
 #[derive(Debug, Clone, Serialize, Deserialize, utoipa::ToSchema)]
 pub struct DnsQueryLogEntry {
