+++ source/daemon/crates/wardnet-common/src/tests/mod.rs
@@ -6,7 +6,9 @@ mod dhcp;
 mod dns;
 mod dns_filter;
 mod event;
+mod net;
 mod routing;
+mod serde_util;
 mod tunnel;
 mod vpn_provider;
 mod wireguard_config;
